Refactor Unzip to create zipReader within method
For a cleaner and more consistent API with untar, create the
zip reader within its public method, not at the caller.
