+++ pkg/archive/untar.go
@@ -1,15 +1,7 @@
-// Copyright 2017 The Go Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style
-// license that can be found in the LICENSE file.
-
-// Edited on 2019-10-11 to remove support for nested folders when un-taring
-// so that all files are placed in the same target directory
-
-package helm
+package archive
 
 import (
 	"archive/tar"
-	"archive/zip"
 	"compress/gzip"
 	"fmt"
 	"io"
@@ -23,86 +15,18 @@ import (
 
 // TODO(bradfitz): this was copied from x/build/cmd/buildlet/buildlet.go
 // but there were some buildlet-specific bits in there, so the code is
-// forked for now.  Unfork and add some opts arguments here, so the
+// forked for now. Unfork and add some opts arguments here, so the
 // buildlet can use this code somehow.
 
 // Untar reads the gzip-compressed tar file from r and writes it into dir.
+// Copyright 2017 The Go Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+//
+// Edited on 2019-10-11 to remove support for nested folders when un-taring
+// so that all files are placed in the same target directory
 func Untar(r io.Reader, dir string) error {
 	return untar(r, dir)
-}
-
-// Untar reads the compressed zip file from r and writes it into dir.
-func Unzip(r zip.Reader, dir string) error {
-	return unzip(r, dir)
-}
-
-func unzip(r zip.Reader, dir string) (err error) {
-	if err != nil {
-		return err
-	}
-	t0 := time.Now()
-	nFiles := 0
-	madeDir := map[string]bool{}
-	defer func() {
-		td := time.Since(t0)
-		if err == nil {
-			log.Printf("extracted zip into %s: %d files, %d dirs (%v)", dir, nFiles, len(madeDir), td)
-		} else {
-			log.Printf("error extracting zip into %s after %d files, %d dirs, %v: %v", dir, nFiles, len(madeDir), td, err)
-		}
-	}()
-
-	// Closure to address file descriptors issue with all the deferred .Close() methods
-	extractAndWriteFile := func(f *zip.File) error {
-		rc, err := f.Open()
-		if err != nil {
-			return err
-		}
-		defer func() {
-			if err := rc.Close(); err != nil {
-				panic(err)
-			}
-		}()
-		baseFile := filepath.Base(f.Name)
-		abs := path.Join(dir, baseFile)
-
-		fmt.Println(abs)
-
-		fi := f.FileInfo()
-		mode := fi.Mode()
-
-		switch {
-		case mode.IsDir():
-			break
-		case mode.IsRegular():
-			f, err := os.OpenFile(abs, os.O_WRONLY|os.O_CREATE|os.O_TRUNC, f.Mode())
-			if err != nil {
-				return err
-			}
-			defer func() {
-				if err := f.Close(); err != nil {
-					panic(err)
-				}
-			}()
-
-			_, err = io.Copy(f, rc)
-			if err != nil {
-				return err
-			}
-		default:
-		}
-		nFiles++
-		return nil
-	}
-
-	for _, f := range r.File {
-		err := extractAndWriteFile(f)
-		if err != nil {
-			return err
-		}
-	}
-
-	return nil
 }
 
 func untar(r io.Reader, dir string) (err error) {
